Fixed inconsistent files

+++ Implementation/Source/vl53l0x_api.c
@@ -391,7 +391,7 @@ VL53L0X_Error VL53L0X_DataInit(VL53L0X_DEV Dev)
 	/* read WHO_AM_I */
 	uint8_t b;
 	Status = VL53L0X_RdByte(Dev, 0xC0, &b);
-	////Serial.print("WHOAMI: 0x"); //Serial.println(b, HEX);
+	//Serial.print("WHOAMI: 0x"); Serial.println(b, HEX);
 	   
 	/* read WHO_AM_I */
 
@@ -2820,14 +2820,14 @@ VL53L0X_Error VL53L0X_GetInterruptThresholds(VL53L0X_DEV Dev,
 
 	Status = VL53L0X_RdWord(Dev, VL53L0X_REG_SYSTEM_THRESH_LOW, &Threshold16);
 	/* Need to multiply by 2 because the FW will apply a x2 */
-	*pThresholdLow = (FixPoint1616_t)(((uint32_t) 0x00fff & Threshold16) << 17);
+	*pThresholdLow = ((FixPoint1616_t)(0x00fff & Threshold16) << 17);
 
 	if (Status == VL53L0X_ERROR_NONE) {
 		Status = VL53L0X_RdWord(Dev, VL53L0X_REG_SYSTEM_THRESH_HIGH,
 			&Threshold16);
 		/* Need to multiply by 2 because the FW will apply a x2 */
 		*pThresholdHigh =
-			(FixPoint1616_t)(((uint32_t)0x00fff & Threshold16) << 17);
+			((FixPoint1616_t)(0x00fff & Threshold16) << 17);
 	}
 
 	LOG_FUNCTION_END(Status);
